REF: move Series-specific methods from NDFrame (#40776)
